+++ tests/test_standalone_stac_item.py
@@ -1,55 +1,45 @@
 import json
 import pytest
-import requests
 import os
 import tempfile
 from urllib.parse import quote
-
-import xncml
+import responses
+from pathlib import Path
 
 from STACpopulator.implementations.CMIP6_UofT.add_CMIP6 import CMIP6ItemProperties, CMIP6populator
 from STACpopulator.input import THREDDSLoader
 from STACpopulator.models import GeoJSONPolygon
-from STACpopulator.stac_utils import STAC_item_from_metadata
+from STACpopulator.stac_utils import STAC_item_from_metadata, ds_attrs
+import pystac
+from conftest import reference_path_from_url, DATA
 
 CUR_DIR = os.path.dirname(__file__)
-
+TEST_NC_URLS = ["https://pavics.ouranos.ca/twitcher/ows/proxy/thredds/birdhouse/testdata/xclim/cmip6"
+                            "/sic_SImon_CCCma-CanESM5_ssp245_r13i1p2f1_2020.nc",
+                ]
 
 def quote_none_safe(url):
     return quote(url, safe="")
 
 
-@pytest.mark.online
-def test_standalone_stac_item_thredds_ncml():
-    thredds_url = "https://pavics.ouranos.ca/twitcher/ows/proxy/thredds"
-    thredds_path = "birdhouse/testdata/xclim/cmip6"
-    thredds_nc = "sic_SImon_CCCma-CanESM5_ssp245_r13i1p2f1_2020.nc"
-    thredds_catalog = f"{thredds_url}/catalog/{thredds_path}/catalog.html"
-    thredds_ds = f"{thredds_path}/{thredds_nc}"
-    thredds_ncml_url = (
-        f"{thredds_url}/ncml/{thredds_path}/{thredds_nc}"
-        f"?catalog={quote_none_safe(thredds_catalog)}&dataset={quote_none_safe(thredds_ds)}"
-    )
-
-    # FIXME: avoid hackish workarounds
-    data = requests.get(thredds_ncml_url).text
-    attrs = xncml.Dataset.from_text(data).to_cf_dict()
-    attrs["access_urls"] = {  # FIXME: all following should be automatically added, but they are not!
-        "HTTPServer": f"{thredds_url}/fileServer/{thredds_path}/{thredds_nc}",
-        "OPENDAP": f"{thredds_url}/dodsC/{thredds_path}/{thredds_nc}",
-        "WCS": f"{thredds_url}/wcs/{thredds_path}/{thredds_nc}?service=WCS&version=1.0.0&request=GetCapabilities",
-        "WMS": f"{thredds_url}/wms/{thredds_path}/{thredds_nc}?service=WMS&version=1.3.0&request=GetCapabilities",
-        "NetcdfSubset": f"{thredds_url}/ncss/{thredds_path}/{thredds_nc}/dataset.html",
-    }
-
+@pytest.mark.parametrize("url", TEST_NC_URLS)
+@responses.activate
+def test_standalone_stac_item_thredds_ncml(url):
+    responses._add_from_file(file_path=DATA / "responses.yaml")
+    attrs = ds_attrs(url)
     stac_item_id = CMIP6populator.make_cmip6_item_id(attrs["attributes"])
     stac_item = STAC_item_from_metadata(stac_item_id, attrs, CMIP6ItemProperties, GeoJSONPolygon)
 
-    ref_file = os.path.join(CUR_DIR, "data/stac_item_testdata_xclim_cmip6_ncml.json")
-    with open(ref_file, mode="r", encoding="utf-8") as ff:
-        reference = json.load(ff)
+    assert isinstance(stac_item, pystac.Item)
 
-    assert stac_item.to_dict() == reference
+    schemas_validated = stac_item.validate()
+    assert len(schemas_validated) >= 1
+    assert "item.json" in schemas_validated[0]
+
+    # Compare with stored item dict
+    path = reference_path_from_url(url)
+    expected = json.loads(path.read_text())
+    assert stac_item.to_dict() == expected
 
 
 class MockedNoSTACUpload(CMIP6populator):